Allow gcc warnings

+++ scripts/gcc-wrapper.py
@@ -64,13 +64,16 @@ def interpret_warning(line):
     if m and m.group(2) not in allowed_warnings:
         print("error, forbidden warning:", m.group(2))
 
-        # If there is a warning, remove any object if it exists.
-        if ofile:
-            try:
-                os.remove(ofile)
-            except OSError:
-                pass
-        sys.exit(1)
+        # We let this pass for now, gcc 10 and clang have much more
+        # warnings than upstream supported compilers...
+        if 0:
+            # If there is a warning, remove any object if it exists.
+            if ofile:
+                try:
+                    os.remove(ofile)
+                except OSError:
+                    pass
+            sys.exit(1)
 
 def run_gcc():
     args = sys.argv[1:]
